Opt out from `RLIMIT_NPROC 0 0`
Example scenario when we encounter this issue was the `sshd`.
The `RLIMIT_NPROC` is set for the `sshd` user and we cannot rely
on approach  adjusting the limit. During client authentication
phase `sshd` is the onwer of the `sshd` process. After the client
successfully authenticate the `sshd` process which
belongs to `sshd` user is gone. The `sshd` process belongs to client
user.

Example scenario when we encounter the issue is when two clients
connects using password instead of public/private key and waits
on screen requested for password input. During this time `sshd`
user own both of the client's `sshd` process.
Since we cannot determine how much process the `sshd` user currently
own we opt out from rlimit call.