Prefix `libunwind` symbols
`libunwind` is a library widely used by applications.
In AppScope, it is used to generate a backtrace in case of a crash.
One side effect of integrating the libunwind library in AppScope (1.3.0) is
that we expose libunwind symbols, such as `unw_backtrace`. The application
can call `unw_backtrace` with its own intentions. When the AppScope library
is loaded, it can result in calling the AppScope variant of `libunwind`
functions, which can result in a SIGSEGV. This patch allows to
prefix `libunwind` functions name which will distinguish the AppScope
`libunwind` symbols.

Fixes: criblio#1499