Modify `__sprintf_chk` call in `wrap.c`
The problematic code resides in `_nc_read_entry2`.

```
char filename[PATH_MAX]
...

    _nc_SPRINTF(filename, _nc_SLIMIT(PATH_MAX)
		"%.*s", PATH_MAX - 1, name);
```

The variable `name` above is set to "xterm".

In the absence of libscope.so, the aforementioned function assigns
the value "xterm" to the filename variable. However, when libscope.so
is loaded, the function assigns "xter" to filename.

In the Fedora environment, the invocation of `__sprintf_chk`
leads to the execution of `__vsprintf_internal`:

```
   va_start (ap, format);
   ret = __vsprintf_internal(s, slen, format, ap, mode);
   va_end (ap);
```

On the other hand, AppScope utilizes the following call:

```
   va_start(ap, format);
   rc = vsnprintf(str, strlen, format, ap);
   va_end(ap);
```

It should be noted that the implementations of `vsnprintf` and
`__vsprintf_internal` differ in standard library. This commit replaces
the `vsnprintf` call with `vsprintf` to address the discrepancy.

Fixes #1567